Styling updated - height: 100vh or not?
If I removed the height: 100vh; I don't have the gap between
header and hero sections, but the footer isn't stuck to the bottom.

I think I want to keep it, but need to remove the gap from header
to hero.

Links need some <a> tags, and the subtext under the info section
icon images spills a bit wide on a wider (desktop) view.